added basic UI + updated just runner
it's not the most polished in the world, but it should get us started to
try out some web socket stuff. a more fleshed out design can come later.

I've also updated a few things on the just runner to work with both
windows and linux. found out pnpm is hard on something like docker.