temp fixed node_modules issue +sql implemented +user model
still trying to figure out .pnpm-store and how to use it in docker but
fixed the build to be working for now. I've also gone ahead and added
the startings of an implementation for mysql and the starts of a model
for the user. It's a bit of a slow going process but it's been fun!